set up typescript, added tsconfig.json file and conigured it, added types directory and a images.d.ts file so that pngs are recognized properly by typescript, and converted map.jsx to .tsx. ONE ISSUE: tailwind on images is no longer recognized in map.tsx, which means the sizing is whack. Working on this more later
