+++ frontend/package.json
@@ -24,7 +24,6 @@
     "react-dom": "^18.2.0",
     "react-icons": "^5.0.1",
     "react-router-dom": "^6.22.1",
-    "react-scripts": "^5.0.1",
     "sass": "^1.71.0",
     "web-vitals": "^2.1.4"
   },
@@ -53,6 +52,10 @@
     ]
   },
   "devDependencies": {
-    "tailwindcss": "^3.4.1"
+    "@types/react": "^18.2.58",
+    "@types/react-dom": "^18.2.19",
+    "react-scripts": "^5.0.1",
+    "tailwindcss": "^3.4.1",
+    "typescript": "^4.9.5"
   }
 }
